- Fixed build_optimized.gradle can't compile.

+++ src/main/java/net/minecraft/entity/item/EntityMinecartFurnace.java
@@ -29,7 +29,7 @@ public class EntityMinecartFurnace extends EntityMinecart {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, (byte) 0);
+        this.dataWatcher.addObject(16, Byte.valueOf((byte) 0));
     }
 
     /**
